Add files via upload
n this module, I have the opportunity to show my Python skills, a must-have in the world of Data Science and Data Analysis. We'll delve into Python fundamentals and data structures while working with data. What's exciting is that we have a hands-on project where I get to step into the shoes of a Data Scientist or Data Analyst. I'll be harnessing the power of Python, leveraging libraries like yfinance and Pandas, and even indulging in a bit of web scraping to collect data. All of this will take place in the friendly environment of a Jupyter notebook, where I'll craft an engaging data dashboard. The ultimate aim? To create a dashboard that provides comprehensive insights into both stock performance and revenue data.